refactor(mb-api): mutualiser agreger dans resolveAgregation + renommer cibles→individusCibles
