+++ apps/mb-api/src/indicateur/resolveAgregation.ts
@@ -1,5 +1,7 @@
 import { type FonctionAgregation } from '@pilote/mb-shared/indicateur'
 
+import { Decimal } from '@/framework/decimal'
+
 // Un individu est agrégé pour cet indicateur s'il a au moins un enfant direct
 // ET que son référentiel est configuré avec une fonction d'agrégation active
 // (≠ 'NONE').
@@ -18,4 +20,21 @@ export const getFonctionAgregationActive = (
   const fonction = ctx.fonctionAgregationParReferentiel.get(referentielId)
   if (!fonction || fonction === 'NONE') return null
   return fonction
+}
+
+// Précondition : `valeurs` est non vide et `fonction` est active (≠ 'NONE').
+// `AVG` est une moyenne arithmétique simple non pondérée.
+export const agreger = (valeurs: ReadonlyArray<Decimal>, fonction: FonctionAgregation): Decimal => {
+  const somme = valeurs.reduce((acc, v) => acc.plus(v), new Decimal(0))
+  switch (fonction) {
+    case 'SUM':
+      return somme
+    case 'AVG':
+      return somme.dividedBy(valeurs.length)
+    case 'NONE':
+      throw new Error(
+        "agreger appelé avec fonction 'NONE' : une dérivation ne devrait pas être déclenchée " +
+          "lorsque l'agrégation est désactivée.",
+      )
+  }
 }
