+++ apps/mb-api/src/valeurAvancement/queries/listValeursForIndicateur.ts
@@ -50,21 +50,27 @@ const buildSeries = async ({
   indicateurPublicId: string
   params: ListValeursForIndicateurQuery
 }): Promise<ValeurAvancementListApiModel> => {
-  const cibles = await loadIndividusParPublicId(params.individus)
-  if (cibles.length === 0) return { items: [] }
+  const individusCibles = await loadIndividusParPublicId(params.individus)
+  if (individusCibles.length === 0) return { items: [] }
 
   const dateTrunc: DateTrunc = params.dateTrunc ?? DEFAULT_DATE_TRUNC
   const startedAt = performance.now()
-  const { ctx, allNodes } = await loadResolveSerieContext({ indicateurId, cibles, dateTrunc })
+  const { ctx, allNodes } = await loadResolveSerieContext({
+    indicateurId,
+    cibles: individusCibles,
+    dateTrunc,
+  })
   const cache = new Map<string, ReadonlyArray<PointInterne>>()
 
   const items: ValeurAvancementApiModel[] = []
-  for (const cible of cibles) {
-    const serie = await resolveSerieIndividu(cible.id, ctx, cache)
+  for (const individuCible of individusCibles) {
+    const serie = await resolveSerieIndividu(individuCible.id, ctx, cache)
     for (const point of serie) {
       if (params.dateDebut && point.bucket < params.dateDebut) continue
       if (params.dateFin && point.bucket > params.dateFin) continue
-      items.push(toApiModel({ indicateurPublicId, individuPublicId: cible.publicId, point }))
+      items.push(
+        toApiModel({ indicateurPublicId, individuPublicId: individuCible.publicId, point }),
+      )
     }
   }
   logger.info(
@@ -72,7 +78,7 @@ const buildSeries = async ({
       event: 'valeurAvancement.listValeursForIndicateur.timing',
       indicateurId,
       dateTrunc,
-      nbCibles: cibles.length,
+      nbCibles: individusCibles.length,
       nbNodes: allNodes.size,
       nbPoints: items.length,
       durationMs: Math.round(performance.now() - startedAt),
